Add CellTrigger support to ProximityCapturable
